// src/app/api/models/page-bc-org-hierarchy-projection-with-out-coordinators.ts
/* tslint:disable */
/* eslint-disable */
import { BcOrgHierarchyProjectionWithOutCoordinators } from './bc-org-hierarchy-projection-with-out-coordinators';
import { PageableObject } from './pageable-object';
import { SortObject } from './sort-object';
export interface PageBcOrgHierarchyProjectionWithOutCoordinators {
  content?: Array<BcOrgHierarchyProjectionWithOutCoordinators>;
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  pageable?: PageableObject;
  size?: number;
  sort?: SortObject;
  totalElements?: number;
  totalPages?: number;
}